requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt[crypto]==5.5.1
django-filter==25.1
drf-spectacular==0.28.0
python-dotenv==1.0.1
dj-database-url==2.3.0
whitenoise==6.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
